bite: fix auction lot size, instead of debt
Collateral auctions were previously fixed in terms of the debt they
needed to cover, but are now fixed in terms of collateral on offer,
i.e. the `lump` is now denominated in collateral, rather than Dai.

If a Dai quantity were to be re-introduced in the future, it should use a
price feed to determine the amount of collateral that is equal to some
fixed size, rather than reasoning about the debt.